chore(background): 🤖 move background to background directory
